[Audio][Train] Reg w/ diff losses; Classification
Purpose of commit: Previous runs with audio valence regression model
was producing sensible highs and lows, but the range was very limited
(~0.45 to ~0.55, or worse) and was converging a mean of 0.5, which
is the mean of the data. This is a relatively common problem with
neural network regression. Found some resources online citing the
problem, but little solutions. Note: the author of the CNN and RCNN
papers for audio classification faced the exact same problem
and had some github questions on keras and on his blog.

The only thing I wanted to try before switching to classification
was using a L1 and Huber loss instead of L2.

Results: Regression still has same problem basically, classification
works better.

Additional (should be a separate commit): refactored MovieReader
and analysis/prepare_data.py --save_video_frames to work with
sample rate < 1, i.e. 0.5. Needed to treat sr as float.
NOTE: filenames are saved as frame_1_0h00m01s.jpg, frame_2_0h00m01s.jpg
for sr=0.5, i.e. second is rounded.

Regression with different loss:
- L1: 2017-02-07___14-56-18
python tasks/audio_sent/main.py -bs 32 --optim adam -lr 0.001 --bn_decay 0.9 --reg_loss L1 --gpus 2 -e 75
python tasks/audio_sent/main.py --mode predict --ckpt_dir 2017-02-07___14-56-18 --load_epoch 3 -vd data/videos/tmp/ --gpus 0 --stride 5 -bs 1
- Huber: 2017-02-07___15-00-39
python tasks/audio_sent/main.py -bs 32 --optim adam -lr 0.001 --bn_decay 0.9 --reg_loss huber -e 75 --gpus 3
python tasks/audio_sent/main.py --mode predict --ckpt_dir 2017-02-07___15-00-39 --load_epoch 7 -vd data/videos/tmp/ --gpus 0 --stride 5 -bs 1

Classification - different approaches (I tried A, went with B)
A) Filter in batch join
- Didn't work because having trouble with variable batch size after items are filtered out
    - in AudioCNN, batch_size is ?
- http://stackoverflow.com/questions/33903569/how-to-filter-tensor-from-queue-based-on-some-predicate-in-tensorflow
- http://stackoverflow.com/questions/35853483/tensorflow-constant-with-variable-size
- Also see previous commit for progressive fine-tuning:
    - https://github.mit.edu/echu/a-darn-good-yarn/commit/6cb1394b7926d9b7e23199c55568755db2c6bd12

Code:
-- Filter to those examples
if self.params['obj'] == 'valence_class':
    neg_match = tf.less(label_batch, [VALENCE_CLASS_THRESHOLD])
    pos_match = tf.greater(label_batch, [1-VALENCE_CLASS_THRESHOLD])
    neg = tf.gather(clip_batch, tf.reshape(tf.where(neg_match), [-1]))
    pos = tf.gather(clip_batch, tf.reshape(tf.where(pos_match), [-1]))
    neg_label = tf.fill([tf.reduce_sum(tf.cast(neg_match, tf.int32))], 0)
    pos_label = tf.fill([tf.reduce_sum(tf.cast(pos_match, tf.int32))], 1)
    clip_batch = tf.concat(0, [neg, pos])
    label_batch = tf.concat(0, [neg_label, pos_label])

*B) Create new tfrecords with valence_class label
- Did this. Read existing tfrecords, save new ones with valence_class label
python tasks/audio_sent/prepare_data.py --write_tfrecords_valence_class
- Test plan:
- Train
    - python tasks/audio_sent/main.py --gpus 2,3 -bs 32 --optim adam  -lr 0.001 -obj valence_class -e 75
- Test
    - python tasks/audio_sent/main.py --mode test --ckpt_dir 2017-02-08___08-53-28 --load_epoch 1 --gpus 0 -bs 32 -obj valence_class
- Predict
    - python tasks/audio_sent/main.py --mode predict --ckpt_dir 2017-02-08___08-53-28 --load_epoch 0 -vd data/videos/tmp/ --gpus 0 --stride 5 -bs 1 -bj valence_class
- Look at GUI, run extract_highlight

C) Once upgraded to tf 1.0, can use tf.train.maybe_shuffle_batch_join and remove above tfrecords

+++ emotion/core/utils/MovieReader.py
@@ -141,4 +141,4 @@ if __name__ == "__main__":
     args = parser.parse_args()
  
     mr = MovieReader()
-    mr.write_frames(args.input_path, args.output_dir, int(args.sample_rate_in_sec), args.target_w, args.target_h)
+    mr.write_frames(args.input_path, args.output_dir, float(args.sample_rate_in_sec), args.target_w, args.target_h)
